+++ Core/CryptoExchanges/Models/MarketFullAggregatedTrade.swift
@@ -4,6 +4,6 @@ public struct MarketFullAggregatedTrade: Codable {
     public let id: Int
     public let date: Date
     public let symbol: String
-    public let price: Decimal
-    public let quantity: Decimal
+    public let price: Double
+    public let quantity: Double
 }
