+++ Core/CryptoExchanges/Models/MarketTicker.swift
@@ -5,8 +5,8 @@ public struct MarketTicker: Codable {
     public var id: Int
     public var date: Date
     public var symbol: String
-    public var bidPrice: Decimal
-    public var bidQuantity: Decimal
-    public var askPrice: Decimal
-    public var askQuantity: Decimal
+    public var bidPrice: Double
+    public var bidQuantity: Double
+    public var askPrice: Double
+    public var askQuantity: Double
 }
